refactor: rename wallet service plugin
