+++ packages/mask/src/extension/popups/pages/Wallet/CreateWallet/index.tsx
@@ -1,15 +1,14 @@
-import { memo, useCallback, useState } from 'react'
-import { Button, Typography } from '@mui/material'
 import { makeStyles } from '@masknet/theme'
+import { Button, Typography } from '@mui/material'
+import { memo, useCallback, useState } from 'react'
 import { Controller } from 'react-hook-form'
-import type { z as zod } from 'zod'
-import { StyledInput } from '../../../components/StyledInput/index.js'
 import { useNavigate } from 'react-router-dom'
-import { WalletRPC } from '../../../../../plugins/Wallet/messages.js'
+import type { z as zod } from 'zod'
+import { WalletRPC } from '../../../../../plugins/WalletService/messages.js'
 import { useI18N } from '../../../../../utils/index.js'
-
-import { useSetWalletNameForm } from '../hooks/useSetWalletNameForm.js'
+import { StyledInput } from '../../../components/StyledInput/index.js'
 import { useTitle } from '../../../hook/useTitle.js'
+import { useSetWalletNameForm } from '../hooks/useSetWalletNameForm.js'
 
 const useStyles = makeStyles()({
     content: {
