+++ packages/mask/src/plugins/WalletService/Worker/index.ts
@@ -1,5 +1,4 @@
 import type { Plugin } from '@masknet/plugin-infra'
-import { base } from '@masknet/plugin-wallet'
 import { MaskMessages } from '@masknet/shared-base'
 import { setWalletBackupProvider } from '../../../../background/services/backup/internal_wallet.js'
 import '../messages.js'
@@ -15,6 +14,7 @@ import {
 } from '../services/index.js'
 import { INTERNAL_getPasswordRequired } from '../services/wallet/password.js'
 import { setupDatabase } from '../database/Plugin.db.js'
+import { base } from '../base.js'
 
 const worker: Plugin.Worker.Definition = {
     ...base,
