+++ packages/mask/src/plugins/WalletService/base.ts
@@ -1,0 +1,13 @@
+import type { Plugin } from '@masknet/plugin-infra'
+import { PLUGIN_DESCRIPTION, PLUGIN_ID, PLUGIN_NAME } from './contants.js'
+
+export const base: Plugin.Shared.Definition = {
+    ID: PLUGIN_ID,
+    name: { fallback: PLUGIN_NAME },
+    description: { fallback: PLUGIN_DESCRIPTION },
+    publisher: { name: { fallback: 'Mask Network' }, link: 'https://mask.io/' },
+    enableRequirement: {
+        networks: { type: 'opt-out', networks: {} },
+        target: 'stable',
+    },
+}
