+++ packages/mask/src/plugins/WalletService/contants.ts
@@ -1,0 +1,7 @@
+import { PluginID } from '@masknet/shared-base'
+
+// #region plugin definitions
+export const PLUGIN_ID = PluginID.WalletService
+export const PLUGIN_NAME = 'Wallet Service'
+export const PLUGIN_DESCRIPTION = 'Mask Wallet Service'
+// #endregion
