+++ packages/mask/src/plugins/WalletService/database/Wallet.db.ts
