+++ packages/mask/src/plugins/WalletService/database/types.ts
