+++ packages/mask/src/plugins/WalletService/services/helpers.ts
