+++ packages/mask/src/plugins/WalletService/services/legacyWallet.ts
