+++ packages/mask/src/plugins/WalletService/services/wallet/password.ts
